Crate bevy_time

Modules

prelude

The Bevy Time Prelude.

Structs

FixedTimestep

A system run criteria that enables systems or stages to run at a fixed timestep between executions.

FixedTimestepState

The internal state of each FixedTimestep.

FixedTimesteps

A global resource that tracks the individual FixedTimestepStates for every labeled FixedTimestep.

Stopwatch

A Stopwatch is a struct that track elapsed time when started.

Time

Tracks elapsed time since the last update and since the App has started

TimePlugin

Adds time functionality to Apps.

TimeReceiver

Channel resource used to receive time from render world

TimeSender

Channel resource used to send time from render world

TimeSystem

Updates the elapsed time. Any system that interacts with Time component should run after this.

Timer

Tracks elapsed time. Enters the finished state once duration is reached.

Functions

create_time_channels

Creates channels used for sending time between render world and app world